Fluxes: merged include path to both KT/collinear fluxes
