+++ CepGenAddOns/APFELWrapper/CollinearFlux.cpp
@@ -20,8 +20,8 @@
 
 #include <cmath>
 
-#include "CepGen/CollinearFluxes/CollinearFlux.h"
 #include "CepGen/Modules/PartonFluxFactory.h"
+#include "CepGen/PartonFluxes/CollinearFlux.h"
 #include "CepGen/Physics/PDG.h"
 #include "CepGen/Utils/Message.h"
