+++ CepGenAddOns/BoostWrapper/PythonObjectsWrappers.h
@@ -1,6 +1,6 @@
 /*
  *  CepGen: a central exclusive processes event generator
- *  Copyright (C) 2023  Laurent Forthomme
+ *  Copyright (C) 2023-2024  Laurent Forthomme
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -21,8 +21,8 @@
 
 #include <boost/python/wrapper.hpp>
 
-#include "CepGen/CollinearFluxes/CollinearFlux.h"
-#include "CepGen/KTFluxes/KTFlux.h"
+#include "CepGen/PartonFluxes/CollinearFlux.h"
+#include "CepGen/PartonFluxes/KTFlux.h"
 
 namespace py = boost::python;
