+++ src/CollinearFluxes/CollinearFlux.cpp
@@ -16,7 +16,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "CepGen/CollinearFluxes/CollinearFlux.h"
+#include "CepGen/PartonFluxes/CollinearFlux.h"
 #include "CepGen/Physics/Utils.h"
 
 namespace cepgen {
