+++ src/CollinearFluxes/EPACollinearFlux.cpp
@@ -17,10 +17,10 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "CepGen/CollinearFluxes/CollinearFlux.h"
 #include "CepGen/FormFactors/Parameterisation.h"
 #include "CepGen/Modules/FormFactorsFactory.h"
 #include "CepGen/Modules/PartonFluxFactory.h"
+#include "CepGen/PartonFluxes/CollinearFlux.h"
 #include "CepGen/Physics/PDG.h"
 #include "CepGen/Physics/Utils.h"
