+++ src/KTFluxes/ElasticKTFluxes.cpp
@@ -20,9 +20,9 @@
 
 #include "CepGen/Core/Exception.h"
 #include "CepGen/FormFactors/Parameterisation.h"
-#include "CepGen/KTFluxes/KTFlux.h"
 #include "CepGen/Modules/FormFactorsFactory.h"
 #include "CepGen/Modules/PartonFluxFactory.h"
+#include "CepGen/PartonFluxes/KTFlux.h"
 #include "CepGen/Physics/PDG.h"
 #include "CepGen/Physics/Utils.h"
