+++ src/KTFluxes/KTFlux.cpp
@@ -1,6 +1,6 @@
 /*
  *  CepGen: a central exclusive processes event generator
- *  Copyright (C) 2023  Laurent Forthomme
+ *  Copyright (C) 2023-2024  Laurent Forthomme
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -16,8 +16,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "CepGen/KTFluxes/KTFlux.h"
-#include "CepGen/Physics/PartonFlux.h"
+#include "CepGen/PartonFluxes/KTFlux.h"
 #include "CepGen/Physics/Utils.h"
 
 namespace cepgen {
