+++ utils/cepgenPlotKTFluxes.cc
@@ -19,9 +19,9 @@
 #include <fstream>
 
 #include "CepGen/Generator.h"
-#include "CepGen/KTFluxes/KTFlux.h"
 #include "CepGen/Modules/DrawerFactory.h"
 #include "CepGen/Modules/PartonFluxFactory.h"
+#include "CepGen/PartonFluxes/KTFlux.h"
 #include "CepGen/Utils/ArgumentsParser.h"
 #include "CepGen/Utils/Drawer.h"
 #include "CepGen/Utils/Graph.h"
